+++ client/package.json
@@ -14,6 +14,7 @@
     "@types/leaflet-rotatedmarker": "^0.2.5",
     "axios": "^1.7.7",
     "boundingbox": "^1.4.0",
+    "date-fns": "^4.1.0",
     "leaflet": "^1.9.4",
     "leaflet-polylinedecorator": "^1.6.0",
     "leaflet-rotatedmarker": "^0.2.0",
@@ -42,8 +43,8 @@
     "globals": "^15.11.0",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.11.0",
-    "vite": "^5.4.10",
-    "vite-node": "^2.1.5",
+    "vite": "^6.3.5",
+    "vite-node": "^3.1.3",
     "vite-plugin-pwa": "^0.21.0"
   }
 }
