+++ client/src/Types/Express.ts
@@ -1,0 +1,8 @@
+import { UserI } from "./User";
+
+export interface AuthResponse {
+  token?: string;
+  message?: string;
+  updated?: UserI;
+  user?: UserI;
+}
