+++ client/src/styles/_Visualisations.css
@@ -1,0 +1,42 @@
+
+.visualisations-page {
+  font-family: Arial, sans-serif;
+  margin: 0;
+  padding: 20px;
+  background-color: #232323;
+  color: white;
+}
+
+h1 {
+  text-align: center;
+  font-size: 2rem;
+  margin-bottom: 20px;
+  color: #ffffff;
+}
+
+h2 {
+  font-size: 1.5rem;
+  margin-bottom: 20px;
+  color: #ffffff; /* Theme color */
+  text-align: center;
+}
+
+
+.section {
+  margin-bottom: 40px;
+  background: #2f2f2f;
+  padding: 20px;
+  border-radius: 10px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.5);
+}
+
+
+.chart-card {
+  margin: 20px auto;
+  background: #1e1e1e;
+  padding: 15px;
+  border-radius: 8px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.5);
+  text-align: center;
+  max-width: 800px;
+}
