+++ server/Types/Share.ts
@@ -3,6 +3,7 @@ import { RouteI } from "./Route.js";
 
 export interface ShareI {
     _id?: string;
+    date: string;
     owner: UserI;
     route: RouteI;
     password: string;
